+++ src/tap/communication/sensors/bmi088/bmi088.cpp
@@ -69,7 +69,7 @@ void Bmi088::requestRecalibration()
     }
 }
 
-void Bmi088::initialize()
+void Bmi088::initialize(float sampleFrequency, float mahonyKp, float mahonyKi)
 {
 #if !defined(PLATFORM_HOSTED)
     ImuCS1Accel::GpioOutput();
@@ -89,6 +89,8 @@ void Bmi088::initialize()
     initializeGyro();
 
     imuHeater.initialize();
+
+    mahonyAlgorithm.begin(sampleFrequency, mahonyKp, mahonyKi);
 }
 
 void Bmi088::initializeAcc()
@@ -178,6 +180,9 @@ void Bmi088::periodicIMUUpdate()
     uint8_t rxBuff[6] = {};
 
     Bmi088Hal::bmi088AccReadMultiReg(Acc::ACC_X_LSB, rxBuff, 6);
+
+    prevIMUDataReceivedTime = tap::arch::clock::getTimeMicroseconds();
+
     data.accRaw[ImuData::X] = bigEndianInt16ToFloat(rxBuff);
     data.accRaw[ImuData::Y] = bigEndianInt16ToFloat(rxBuff + 2);
     data.accRaw[ImuData::Z] = bigEndianInt16ToFloat(rxBuff + 4);
@@ -244,7 +249,7 @@ void Bmi088::computeOffsets()
         data.accOffsetRaw[ImuData::Y] /= BMI088_OFFSET_SAMPLES;
         data.accOffsetRaw[ImuData::Z] /= BMI088_OFFSET_SAMPLES;
         imuState = ImuState::IMU_CALIBRATED;
-        mahonyAlgorithm = Mahony();
+        mahonyAlgorithm.reset();
     }
 }