+++ src/tap/communication/serial/ref_serial_data.hpp
@@ -26,7 +26,9 @@
 
 #include "modm/architecture/utils.hpp"
 
-namespace tap::serial
+#include "dji_serial.hpp"
+
+namespace tap::communication::serial
 {
 /**
  * Contains enum and struct definitions used in the `RefSerial` class.
@@ -69,7 +71,7 @@ public:
     {
     public:
         RobotToRobotMessageHandler() {}
-        virtual void operator()(const DJISerial::SerialMessage &message) = 0;
+        virtual void operator()(const DJISerial::ReceivedSerialMessage &message) = 0;
     };
 
     /**
@@ -225,6 +227,9 @@ public:
                                           /// in RMUC.
             float bulletSpeed;            /// Last bullet speed (in m/s).
             float yaw;                    /// Barrel yaw position (degree).
+            uint32_t lastReceivedLaunchingInfoTimestamp;  /// Last time in milliseconds that the
+                                                          /// real-time launching information
+                                                          /// message was received
         };
 
         struct RobotData
@@ -304,17 +309,6 @@ public:
         };
 
         /**
-         * All referee data will be prefixed by a header of this structure
-         */
-        struct FrameHeader
-        {
-            uint8_t SOF;  /// Also known as the "content ID" in the ref system protocol appendix
-            uint16_t dataLength;
-            uint8_t seq;
-            uint8_t CRC8;
-        } modm_packed;
-
-        /**
          * Each graphic message has a graphic header inside of the message data.
          */
         struct InteractiveHeader
@@ -350,7 +344,7 @@ public:
 
         struct DeleteGraphicLayerMessage
         {
-            FrameHeader frameHeader;
+            DJISerial::FrameHeader frameHeader;
             uint16_t cmdId;
             InteractiveHeader interactiveHeader;
             uint8_t deleteOperation;
@@ -360,7 +354,7 @@ public:
 
         struct Graphic1Message
         {
-            FrameHeader frameHeader;
+            DJISerial::FrameHeader frameHeader;
             uint16_t cmdId;
             InteractiveHeader interactiveHeader;
             GraphicData graphicData;
@@ -369,7 +363,7 @@ public:
 
         struct RobotToRobotMessage
         {
-            FrameHeader frameHeader;
+            DJISerial::FrameHeader frameHeader;
             uint16_t cmdId;
             InteractiveHeader interactiveHeader;
             uint8_t dataAndCRC16[115];
@@ -377,7 +371,7 @@ public:
 
         struct Graphic2Message
         {
-            FrameHeader frameHeader;
+            DJISerial::FrameHeader frameHeader;
             uint16_t cmdId;
             InteractiveHeader interactiveHeader;
             GraphicData graphicData[2];
@@ -386,7 +380,7 @@ public:
 
         struct Graphic5Message
         {
-            FrameHeader frameHeader;
+            DJISerial::FrameHeader frameHeader;
             uint16_t cmdId;
             InteractiveHeader interactiveHeader;
             GraphicData graphicData[5];
@@ -395,7 +389,7 @@ public:
 
         struct Graphic7Message
         {
-            FrameHeader frameHeader;
+            DJISerial::FrameHeader frameHeader;
             uint16_t cmdId;
             InteractiveHeader interactiveHeader;
             GraphicData graphicData[7];
@@ -404,7 +398,7 @@ public:
 
         struct GraphicCharacterMessage
         {
-            FrameHeader frameHeader;
+            DJISerial::FrameHeader frameHeader;
             uint16_t cmdId;
             InteractiveHeader interactiveHeader;
             GraphicData graphicData;
@@ -459,6 +453,6 @@ inline RefSerialData::RobotId operator-(RefSerialData::RobotId id1, RefSerialDat
     return static_cast<RefSerialData::RobotId>(
         static_cast<uint16_t>(id1) - static_cast<uint16_t>(id2));
 }
-}  // namespace tap::serial
+}  // namespace tap::communication::serial
 
 #endif  // REF_SERIAL_DATA_HPP_